Moved FitWindow into sidebar
fixed bug of toolbar being duplicated when the color scheme is changed,
changed the created of the label and entry widgets from the exec ... method to a dictionary, which can be accessed more easily.
fixed bug of overlapping settings in the sidebar by hardcoding the "self.row" value. Could be improved!