simplified settings frame
introduced init_val for entry and optionMenu
renamed createOptionMenu to createMenu
renamed set_plot_settings to apply_settings
custom ratios of window sizes by specifying height and width

fixed a bug, where the lineout functionality was broken because the return value of plot() was wrong. It should be a line not the tuple containing the line and the error lines!